+++ Sources/Module/Channel/GroupChannel/SBUGroupChannelModule.Input.swift
@@ -188,14 +188,8 @@ extension SBUGroupChannelModule {
                 let originalImage = info[.originalImage] as? UIImage
                 // TODO: need to improved. (mimetype)
                 // for Camera capture
-                guard let image = originalImage?
-                        .fixedOrientation()
-                        .resize(with: SBUGlobals.imageResizingSize) else { return }
-                
-                let imageData = image.jpegData(
-                    compressionQuality: SBUGlobals.isImageCompressionEnabled ?
-                    SBUGlobals.imageCompressionRate : 1.0
-                )
+                guard let image = originalImage?.fixedOrientation(),
+                      let imageData = image.sbu_convertToData() else { return }
                 
                 self.delegate?.groupChannelModule(
                     self,
@@ -226,15 +220,9 @@ extension SBUGroupChannelModule {
                     )
                 default:
                     let originalImage = info[.originalImage] as? UIImage
-                    guard let image = originalImage?
-                            .fixedOrientation()
-                            .resize(with: SBUGlobals.imageResizingSize) else { return }
-                    
-                    let imageData = image.jpegData(
-                        compressionQuality: SBUGlobals.isImageCompressionEnabled ?
-                        SBUGlobals.imageCompressionRate : 1.0
-                    )
-                    
+                    guard let image = originalImage?.fixedOrientation(),
+                          let imageData = image.sbu_convertToData() else { return }
+                
                     self.delegate?.groupChannelModule(
                         self,
                         didPickFileData: imageData,
@@ -278,14 +266,9 @@ extension SBUGroupChannelModule {
                         guard let imageURL = url as? URL else { return }
                         guard let mimeType = SBUUtils.getMimeType(url: imageURL) else { return }
                         
-                        let image = originalImage
+                        guard let imageData = originalImage
                             .fixedOrientation()
-                            .resize(with: SBUGlobals.imageResizingSize)
-                        let imageData = image.jpegData(
-                            compressionQuality: SBUGlobals.isImageCompressionEnabled
-                            ? SBUGlobals.imageCompressionRate
-                            : 1.0
-                        )
+                            .sbu_convertToData() else { return }
                         
                         let fileExtension = imageURL.pathExtension
                         let fileName = "\(Date().sbu_toString(dateFormat: SBUDateFormatSet.Message.fileNameFormat, localizedFormat: false)).\(fileExtension)"