+++ Sources/Module/Channel/OpenChannel/SBUOpenChannelModule.Input.swift
@@ -86,14 +86,8 @@ extension SBUOpenChannelModule {
             guard let imageURL = tempImageURL else {
                 let originalImage = info[.originalImage] as? UIImage
                 // for Camera capture
-                guard let image = originalImage?
-                        .fixedOrientation()
-                        .resize(with: SBUGlobals.imageResizingSize) else { return }
-                
-                let imageData = image.jpegData(
-                    compressionQuality: SBUGlobals.isImageCompressionEnabled ?
-                    SBUGlobals.imageCompressionRate : 1.0
-                )
+                guard let image = originalImage?.fixedOrientation(),
+                      let imageData = image.sbu_convertToData() else { return }
                 
                 self.delegate?.openChannelModule(
                     self,
@@ -123,14 +117,8 @@ extension SBUOpenChannelModule {
                     
                 default:
                     let originalImage = info[.originalImage] as? UIImage
-                    guard let image = originalImage?
-                            .fixedOrientation()
-                            .resize(with: SBUGlobals.imageResizingSize) else { return }
-                    
-                    let imageData = image.jpegData(
-                        compressionQuality: SBUGlobals.isImageCompressionEnabled ?
-                        SBUGlobals.imageCompressionRate : 1.0
-                    )
+                    guard let image = originalImage?.fixedOrientation(),
+                          let imageData = image.sbu_convertToData() else { return }
                     
                     self.delegate?.openChannelModule(
                         self,
@@ -170,14 +158,9 @@ extension SBUOpenChannelModule {
                         guard let imageURL = url as? URL else { return }
                         guard let mimeType = SBUUtils.getMimeType(url: imageURL) else { return }
                         
-                        let image = originalImage
+                        guard let imageData = originalImage
                             .fixedOrientation()
-                            .resize(with: SBUGlobals.imageResizingSize)
-                        let imageData = image.jpegData(
-                            compressionQuality: SBUGlobals.isImageCompressionEnabled
-                            ? SBUGlobals.imageCompressionRate
-                            : 1.0
-                        )
+                            .sbu_convertToData() else { return }
                         
                         let fileExtension = imageURL.pathExtension
                         let fileName = "\(Date().sbu_toString(dateFormat: SBUDateFormatSet.Message.fileNameFormat, localizedFormat: false)).\(fileExtension)"